Consolidate all DAOs into one Database and one Repository to contain all Entities (instead of one Database and Repository per DAO) in a new QueryAppRepository. All method signatures of the old XRepository functions are the same...just consolidated into one Repository. Updated to database version 5. Getting all confusion marks for a class is now supported
