+++ app/src/androidTest/java/com/example/conwayying/query/ConfusionMarkDatabaseTest.java
@@ -5,9 +5,7 @@ import android.content.Context;
 import android.support.test.InstrumentationRegistry;
 import android.support.test.runner.AndroidJUnit4;
 
-import com.example.conwayying.query.data.AcademicClassDatabase;
-import com.example.conwayying.query.data.ConfusionMarkDatabase;
-import com.example.conwayying.query.data.LectureDatabase;
+import com.example.conwayying.query.data.QueryAppDatabase;
 import com.example.conwayying.query.data.entity.AcademicClass;
 import com.example.conwayying.query.data.entity.ConfusionMark;
 import com.example.conwayying.query.data.entity.ConfusionMarkDao;
@@ -29,37 +27,30 @@ import java.util.List;
 @RunWith(AndroidJUnit4.class)
 public class ConfusionMarkDatabaseTest {
 
-    private AcademicClassDatabase academicClassDB;
-    private LectureDatabase lectureDB;
-    private ConfusionMarkDatabase confusionMarkDB;
+    private QueryAppDatabase queryAppDB;
 
     @Before
     public void createDb() {
         Context context = InstrumentationRegistry.getContext();
-        academicClassDB = Room.inMemoryDatabaseBuilder(context, AcademicClassDatabase.class).build();
-        lectureDB = Room.inMemoryDatabaseBuilder(context, LectureDatabase.class).build();
-        confusionMarkDB = Room.inMemoryDatabaseBuilder(context, ConfusionMarkDatabase.class).build();
+        queryAppDB = Room.inMemoryDatabaseBuilder(context, QueryAppDatabase.class).build();
     }
 
     @After
     public void closeDb() throws IOException {
-        academicClassDB.close();
-        lectureDB.close();
-        confusionMarkDB.close();
+        queryAppDB.close();
     }
 
     @Test
     public void testConfusionMarks() throws Exception {
 
         // Arrange
-        // TODO: move this to private function
 
         // Create two classes with 3 lectures each
         AcademicClass cs465 = new AcademicClass("CS 465");
         AcademicClass cs357 = new AcademicClass("CS 357");
 
-        Long cs465Id = academicClassDB.getAcademicClassDao().insert(cs465);
-        Long cs357Id = academicClassDB.getAcademicClassDao().insert(cs357);
+        Long cs465Id = queryAppDB.getAcademicClassDao().insert(cs465);
+        Long cs357Id = queryAppDB.getAcademicClassDao().insert(cs357);
 
 
         List<Date> testDates = new ArrayList<>();
@@ -76,7 +67,7 @@ public class ConfusionMarkDatabaseTest {
         // No lambda expressions in Java => Sad Mike
         List<Long> cs357LectureIds = new ArrayList<>();
         for(int i = 0; i < 3; i++){
-            Long id = lectureDB.getLectureDao().insert(new Lecture(testDates.get(i), cs357Id.intValue()));
+            Long id = queryAppDB.getLectureDao().insert(new Lecture(testDates.get(i), cs357Id.intValue()));
             cs357LectureIds.add(id);
         }
 
@@ -84,12 +75,12 @@ public class ConfusionMarkDatabaseTest {
         Long secondLectureId = cs357LectureIds.get(1);
 
 
-        Long cs465LectureId = lectureDB.getLectureDao().insert(new Lecture(testDates.get(0), cs465Id.intValue()));
+        Long cs465LectureId = queryAppDB.getLectureDao().insert(new Lecture(testDates.get(0), cs465Id.intValue()));
 
 
 
         // Act
-        ConfusionMarkDao confusionMarkDao = confusionMarkDB.getConfusionMarkDao();
+        ConfusionMarkDao confusionMarkDao = queryAppDB.getConfusionMarkDao();
 
         // The first CS 357 lecture gets 3 confusion marks
         for (int i = 0; i < 3; i++){
@@ -115,12 +106,9 @@ public class ConfusionMarkDatabaseTest {
         Assert.assertEquals(7, confusionMarkDao.getAllConfusionMarksForLecture(cs465LectureId.intValue()).size());
         Assert.assertEquals(0, confusionMarkDao.getAllConfusionMarksForLecture(Integer.MAX_VALUE - 1).size());
 
-        // Per-class asserts
-        // TODO: uncomment these tests after figuring out the Room problem with the subquery
-        /*
+        // Per-AcademicClass asserts
         Assert.assertEquals(2 + 3, confusionMarkDao.getAllConfusionMarksForClass(cs357Id.intValue()).size());
         Assert.assertEquals(7, confusionMarkDao.getAllConfusionMarksForClass(cs465Id.intValue()).size());
         Assert.assertEquals(0, confusionMarkDao.getAllConfusionMarksForClass(Integer.MAX_VALUE - 1).size());
-        */
     }
 }
